refactor(shadcnui-blocks): rename NumberAndSecondaryStatWidget to NumberAndSecondaryStat
