+++ libs/shadcnui-blocks/src/index.ts
@@ -1,5 +1,5 @@
 export * from './lib/components/index';
 export * from './lib/components/features-comparison/index';
 export * from './lib/components/pricing-tiers/index';
-export * from './lib/components/number-and-secondary-stat-widget/NumberAndSecondaryStatWidget';
+export * from './lib/components/number-and-secondary-stat/NumberAndSecondaryStat';
 export * from './lib/components/charts/SparkLine';
